requirements.txt
fastapi
pytest-cov>=5.0.0
httpx
uvicorn
requests
llama-cpp-python>=0.2.90 # fallback local; wheel padrão é CPU-only
